\name{SQUIC-package}
\alias{SQUIC-package}
\alias{SQUIC}
\docType{package}
\title{
\packageTitle{SQUIC}
}
\description{
\packageDescription{SQUIC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SQUIC}
\packageIndices{SQUIC}
SQUIC stand for Sparse Quadradic Inverese covarnce matrix estimation ... balh.
}
\author{
Bob
\packageAuthor{SQUIC}

Maintainer: Jack \packageMaintainer{SQUIC}
}
\references{
Large-scale Sparse Inverse Covariance Matrix Estimation, M.Bollhöfer, A. Eftekhari, S. Scheidegger, O. Schenk \url{https://epubs.siam.org/doi/abs/10.1137/17M1147615?journalCode=sjoce3}

Large-scale Sparse Inverse Covariance Matrix Estimation, M.Bollhöfer, A. Eftekhari, S. Scheidegger, O. Schenk \url{https://epubs.siam.org/doi/abs/10.1137/17M1147615?journalCode=sjoce3}

Large-scale Sparse Inverse Covariance Matrix Estimation, M.Bollhöfer, A. Eftekhari, S. Scheidegger, O. Schenk \url{https://epubs.siam.org/doi/abs/10.1137/17M1147615?journalCode=sjoce3}


}
Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
documentation directory ~~
\keyword{ package }
\seealso{
Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
simple examples of the most important functions
}
