\name{SQUIC-package}
\alias{SQUIC-package}
\alias{SQUIC}
\docType{package}
\title{
\packageTitle{SQUIC}
}
\description{
\packageDescription{SQUIC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SQUIC}
\packageIndices{SQUIC}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SQUIC}

Maintainer: \packageMaintainer{SQUIC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
